[BE] fix: id반환 타입 수정, 양도 게시글 양식 수정

+++ backend/business-service/src/main/java/com/footprints/businessservice/app/domain/board/article/api/ArticleController.java
@@ -43,7 +43,7 @@ public class ArticleController {
             throw new ArticleException(ArticleExceptionType.FILE_IS_NOT_IMAGE);
         }
 
-        Integer count = articleService.saveArticle(memberId, request, multipartFiles);
+        Long count = articleService.saveArticle(memberId, request, multipartFiles);
         return ResponseEntity.status(HttpStatus.CREATED).body(new DataResponse<>(count));
     }
